(PC-29847) refactor(contentful): Rename helper
