+++ src/features/home/components/modules/OffersModule.tsx
@@ -6,8 +6,8 @@ import { ModuleData, OffersModule as OffersModuleType } from 'features/home/type
 import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/helpers'
 import { useAdaptOffersPlaylistParameters } from 'libs/algolia/fetchAlgolia/fetchMultipleOffers/helpers/useAdaptOffersPlaylistParameters'
 import { analytics } from 'libs/analytics'
-import { usePlaylistItemDimensionsFromLayout } from 'libs/contentful/dimensions'
 import { ContentTypes } from 'libs/contentful/types'
+import { usePlaylistItemDimensionsFromLayout } from 'libs/contentful/usePlaylistItemDimensionsFromLayout'
 import useFunctionOnce from 'libs/hooks/useFunctionOnce'
 import { useLocation } from 'libs/location/LocationWrapper'
 import { formatDates } from 'libs/parsers/formatDates'
