+++ src/features/home/components/modules/RecommendationModule.tsx
@@ -5,8 +5,8 @@ import { useHomeRecommendedOffers } from 'features/home/api/useHomeRecommendedOf
 import { HomeOfferTile } from 'features/home/components/HomeOfferTile'
 import { RecommendedOffersModule } from 'features/home/types'
 import { analytics } from 'libs/analytics'
-import { usePlaylistItemDimensionsFromLayout } from 'libs/contentful/dimensions'
 import { ContentTypes, DisplayParametersFields } from 'libs/contentful/types'
+import { usePlaylistItemDimensionsFromLayout } from 'libs/contentful/usePlaylistItemDimensionsFromLayout'
 import useFunctionOnce from 'libs/hooks/useFunctionOnce'
 import { useLocation } from 'libs/location/LocationWrapper'
 import { formatDates } from 'libs/parsers/formatDates'
